# gateway/cmd/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Volha gateway REST API",
        "contact": {
            "name": "Alex \"bustard\" Provor",
            "url": "https://breezy.su",
            "email": "help@breezy.su"
        },
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/color/create": {
            "post": {
                "description": "Добавляет новый цвет",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "color"
                ],
                "summary": "Создать цвет",
                "parameters": [
                    {
                        "description": "Новый цвет",
                        "name": "color",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/views.Color"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Цвет успешно создан",
                        "schema": {
                            "$ref": "#/definitions/views.SWGSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат данных",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка на сервере",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Ошибка взаимодействия с сервисом",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/color/delete": {
            "delete": {
                "description": "Удаляет цвет по ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "color"
                ],
                "summary": "Удалить цвет",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID цвета",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Цвет успешно удалён",
                        "schema": {
                            "$ref": "#/definitions/views.SWGSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка на сервере",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Ошибка взаимодействия с сервисом",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/color/getall": {
            "get": {
                "description": "Возвращает список всех доступных цветов",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "color"
                ],
                "summary": "Получить все цвета",
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "schema": {
                            "$ref": "#/definitions/views.SWGColorListResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка на сервере",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Ошибка взаимодействия с сервисом",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/color/update": {
            "put": {
                "description": "Обновляет данные существующего цвета",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "color"
                ],
                "summary": "Обновить цвет",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID цвета",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Обновлённый цвет",
                        "name": "color",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/views.Color"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Цвет успешно обновлён",
                        "schema": {
                            "$ref": "#/definitions/views.SWGSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный ID или данные",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка на сервере",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Ошибка взаимодействия с сервисом",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "views.Color": {
            "type": "object",
            "properties": {
                "hex": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "views.SWGColorListResponse": {
            "type": "object",
            "properties": {
                "colors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/views.Color"
                    }
                }
            }
        },
...
        "views.SWGErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "something went wrong"
                }
            }
        },
...
        "views.SWGSuccessResponse": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string",
                    "example": "operation successful"
                }
            }
        }
    }
}